% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser-summary.R
\name{read_g09output}
\alias{read_g09output}
\title{extract the summary area from gaussian09 log file}
\usage{
read_g09output(file)
}
\arguments{
\item{file}{file path}
}
\value{
summary list
}
\description{
extract the summary area from gaussian09 log file
}
\examples{
library(dplyr)
dir(".") \%>\%
 grep(".out", ., value = TRUE) \%>\%
  lapply(
    function(x){
      res <- read_g09output(x)
      tryCatch(
        c(res$Title ,res$Summary$`CCSD(T)`),
        error = function(e) NA
      )
    }
  )



}
